pyproject.toml
[project]
name = "src"
version = "1.0.0"
description = "Template para APIs utilizando o framework FastAPI e banco de dados Postgres"
authors = [
    {name = "Seu nome",email = "seuemail@email.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.11.0"
factory-boy = "^3.3.3"
testcontainers = "^4.9.2"
freezegun = "^1.5.1"
dependency-injector = "^4.46.0"
loguru = "^0.7.3"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['E402'] 

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint.pylint]
max-args = 10

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'python src/app.py'
pre_test = 'task lint'
test = 'pytest -x --cov-report=html -vv --showlocals --tb=short'
post_test = 'coverage html'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]